Added primitive priority based scheduling
This just sorts threads in the ready queue.
To implement a true priority based scheduler,
we need to modify the functions that modify
priority for threads, and those that create
threads, to make them insert in order, and to
interrupt running of a thread on modification of
priority if a higher thread is available.